Clean interaction between html and listen
Implement a new signal between the html communication file and cordiebot_listen so that fields will be blanked out when data is not returned.  Affects the get by number, get by date, and get all requests.

Also add the uuid field to the PubNub setup and a timestamp the the console messages.